Update README.md
Feature overview: This sketch uses an Arduino Uno or Nano to emulate the multi-pin sequential logic chips found in alarm clock radios of the 1970s and 1980s and drive a corresponding simple display based on the time and button presses by the user.
Parallel inputs: The Arduino can sense locking buttons for time display, alarm-time display, sleep duration display and bright/dark settings assuming buttons which lock and can ground corresponding Arduino pins to 0v. This changes the display accordingly. 
Non-locking buttons can advance the time fast/slow. An additional non-locking button is required to turn off the alarm. These buttons all use separate input pins on the Arduino. 
Parallel output: The Arduino toggles an output pin high/low in order to turn the radio on/off. 
Peripherals required: The Arduino uses the Adafruit 0.56" 7-segment LED display and an RTC DS1307 real-time clock board to provide the time source and display it using LEDs. I2C connections are used for these peripherals.
Clock features: The sketch provides display of hour and minutes only, though seconds are used internally and not displayed. Date display is available from the RTC, but not used.
The alarm function is provided in the sketch instead of using methods provided by the RTC unit which is only used to keep time. A countdown feature allows a "sleep" function to countdown from 60 minutes with the radio driven "on" until it gets to zero. The display has bright/dim features. 
Sketch operation: The main loop runs continuously to sense button states. A timer 'ticks' the clock forward autonomously every second, so the sketch itself acts as an autonomous clock. 
At the top of the hour the sketch gets the time from the RTC unit to correct the sketch clock if needed.
Interrupts are not used. Instead, button presses will activate "if" loops to, e.g., change the display. Thus the button presses merely divert the main loop through "if" loops.
Additional circuitry: The sketch assumes an Arduino driven by a line power supply, e.g. 5v. It is not optimised for battery-driven applications. The output of 5v will require additional components, e.g. a transistor circuit, to activate a radio or other device according to that output state.